feat: use `this` to provide abort signal
BREAKING CHANGE: the abort signal is now provided through `this`

Previously abort signal was provided through arguments and while it
worked for the most part, it made typings more complicated and required
you to use both arguments in getter method definition, even if didn't
want to use the signal. With *this* change, the signal is now provided
through object that is bound to `this` on getter calls.

To migrate, simply replace all usages of `signal` to `this.signal`, as
well as remove all usages of AsyncComputedOptionsGetterOnly type: it is
removed and merged together into AsyncComputedOptions. If you don't use
`watch` method or it does not return anything, you can ignore the
`source` argument passed to the getter method.