feat: instantly resolve when getter is not async
If getter does not return a thenable, then it only makes sense to
instantly resolve with that value.

This is not considered a breaking change since if the API was used
correctly, this should not cause any problems. However, it's now fully
allowed for asyncComputed getters to act like regular computed
references with extra steps, loss of auto-expanding functionality and
unnecessary performance loss (don't do this unless you really need to).