fix!: no longer uses AbortController
BREAKING CHANGE: this removes this.signal in getters and changes them to
onCancel method that accepts callback to be called, you can also query
this.canceled to see if the call was cancelled.

While AbortController is precisely where you'd use that API, oddly it
seems to be unavailable in some environments, therefore it has been
replaced with the custom callbacks implementation.

This unfortunately badly affects minified size but it is what it is, I
guess =(